\name{distmirr}
\alias{distmirr}

\title{Internal function to convert a distance structure to a matrix}
\description{
  For internal use only.
}
\usage{
distmirr(dis)
}

\arguments{
  \item{dis}{a distance structure as computed by \code{\link{dist}}}
}

\value{
  A symmetric distance matrix is returned.
}

 
\seealso{\code{\link{dist}}}
\keyword{internal}
