\name{.dmvnorm}
\alias{.dmvnorm}
\title{Density of a Multivariate Normal Distribution}
\description{Intended for internal use only.}
\usage{
.dmvnorm(x, mu = NA, inv.sigma = NA, logDetCov = NA)
}
\arguments{
  \item{x}{}
  \item{mu}{}
  \item{inv.sigma}{}
  \item{logDetCov}{}
}
\keyword{internal}
